refactor and changes for mSQL
