+++ tests/test_code_distribution_query.py
@@ -1,5 +1,5 @@
 import pytest
-from hutch_bunny.core.query_solvers import DistributionQuery, solve_distribution
+from hutch_bunny.core.solvers.query_solvers import DistributionQuery, solve_distribution
 from hutch_bunny.core.db_manager import SyncDBManager
 from hutch_bunny.core.rquest_dto.result import RquestResult
 from hutch_bunny.core.rquest_dto.file import File
